requirements-web.txt
# Core project dependencies
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0
anthropic>=0.5.0
google-generativeai>=0.3.0
numpy>=1.24.0
scikit-learn>=1.3.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# Web backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# File processing
aiofiles>=23.2.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0 